Resolve "Fix the warnings when compiling with 'gfortran -Wall -Wextra'" part 1

See merge request atmospheric-data-assimilation/midas!996

+++ src/modules/climatologies_mod.f90
@@ -831,7 +831,7 @@ contains
     character(len=*), intent(in) :: varName   ! Variable name
     integer, intent(in) :: iyear  ! Year for required scaling factor
     ! Result:
-    integer :: scaleFactor        ! Units scaling factor
+    real(8) :: scaleFactor        ! Units scaling factor
 
     ! Locals:
     integer, parameter :: numSpecies=5
